+++ assets/kenney-sketch-desert.tsx
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<tileset version="1.10" tiledversion="1.10.2" name="kenney-sketch-desert" tilewidth="256" tileheight="352" tilecount="5" columns="0">
+ <grid orientation="isometric" width="1" height="1"/>
+ <tile id="0">
+  <image width="256" height="352" source="tiles/kenney-sketch-desert/building_dark_sides_windows_E.png"/>
+ </tile>
+ <tile id="1">
+  <image width="256" height="352" source="tiles/kenney-sketch-desert/dirt_center_E.png"/>
+ </tile>
+ <tile id="2">
+  <image width="256" height="352" source="tiles/kenney-sketch-desert/grass_center_S.png"/>
+ </tile>
+ <tile id="3">
+  <image width="256" height="352" source="tiles/kenney-sketch-desert/stairs_full_N.png"/>
+ </tile>
+ <tile id="4">
+  <image width="256" height="352" source="tiles/kenney-sketch-desert/trees_N.png"/>
+ </tile>
+</tileset>